+++ rollup.config.js
@@ -49,7 +49,6 @@ export default [
         output: {
           file: "build/melonjs.js",
           banner: license,
-          footer : "me.deprecated.apply();",
           name : "me",
           format: "umd",
           generatedCode: "es5"