+++ src/lang/deprecated.js
@@ -1,15 +1,13 @@
+import device from "./../system/device.js";
+import { requestPointerLock, exitPointerLock } from "./../input/input.js";
+
 /**
  * placeholder for all deprecated classes and corresponding alias for backward compatibility
- * @namespace deprecated
- * @memberof me
  */
 
 /**
  * display a deprecation warning in the console
- * @public
- * @function
- * @memberof me.deprecated
- * @name warning
+ * @ignore
  * @param {string} deprecated deprecated class,function or property name
  * @param {string} replacement the replacement class, function, or property name
  * @param {string} version the version since when the lass,function or property is deprecated
@@ -42,17 +40,30 @@ export function warning(deprecated, replacement, version) {
     if (console.groupCollapsed) {
         console.groupEnd();
     }
-
-
 };
 
+/**
+ * @public
+ * @type {Function}
+ * @name turnOnPointerLock
+ * @memberof me.device
+ * @deprecated since 10.3.0
+ * @see me.input.requestPointerLock
+ */
+device.turnOnPointerLock = function () {
+    warning("me.device.turnOnPointerLock()", "me.input.requestPointerLock()", "10.3.0");
+    return requestPointerLock();
+};
 
 /**
- * Backward compatibility for deprecated method or properties are automatically
- * applied when automatically generating an UMD bundle (which is the default since version 9.0).
- * @memberof me.deprecated
- * @function apply
+ * @public
+ * @type {Function}
+ * @name turnOffPointerLock
+ * @memberof me.device
+ * @deprecated since 10.3.0
+ * @see me.input.exitPointerLock
  */
- export function apply() {
-     ;
- };
+device.turnOffPointerLock = function () {
+    warning("me.device.turnOffPointerLock()", "me.input.exitPointerLock()", "10.3.0");
+    return exitPointerLock();
+};